Extract test utils to library
