+++ imageloading-core/src/main/java/dev/chrisbanes/accompanist/imageloading/ImageLoadState.kt
@@ -51,6 +51,6 @@ sealed class ImageLoadState {
      */
     data class Error(
         val painter: Painter? = null,
-        val throwable: Throwable? = null
+        val throwable: Throwable
     ) : ImageLoadState()
 }
