+++ imageloading-core/src/main/java/dev/chrisbanes/accompanist/imageloading/RequestActor.kt
@@ -85,7 +85,7 @@ internal class RequestActor<Param, Result>(
     /**
      * Send an input for processing by the [RequestActor].
      */
-    fun send(input: Any?) {
+    fun send(input: Param) {
         channel.offer(input)
     }
 }
