+++ imageloading-testutils/src/main/res/drawable-nodpi/blue_rectangle.png
