+++ imageloading-testutils/src/main/res/drawable-nodpi/red_rectangle.png
