+++ imageloading-testutils/src/main/res/raw/sample.jpg
