--- mdc-theme/README.md
@@ -1,8 +1,0 @@
-# Jetpack Compose + Material Design Components theme
-
-This library has been migrated into the [MDC-Android Compose Theme Adapter](https://material-components.github.io/material-components-android-compose-theme-adapter/) library.
-It contains the exact same functionality and structure, so should hopefully be an easy migration. 🤞
-
-No further updates will be made to the `dev.chrisbanes.accompanist:accompanist-mdc-theme` Maven artifacts, and the source has been removed (still in the history though).
-
-Thanks to anyone who tried this out while we explored what a solution should look like. 🙏
